feat: moved _norm and _get_paulis to utils
